timeout of test cases not specified by static variable but by getTimeout() method

+++ trunk/source/UltimateTest/src/de/uni_freiburg/informatik/ultimatetest/evals/LTLChecker.java
@@ -80,8 +80,11 @@ public class LTLChecker extends AbstractEvaluationTestSuite {
 		// return super.getDirectories();
 	}
 
+	/**
+	 * {@inheritDoc}
+	 */
 	@Override
-	protected int getTimeout() {
+	protected long getTimeout() {
 //		return 180 * 1000;
 		return 3 * 60 * 1000;
 	}
