+++ trunk/source/UltimateTest/src/de/uni_freiburg/informatik/ultimatetest/evals/TACAS2015AlgoComparison.java
@@ -25,8 +25,11 @@ public class TACAS2015AlgoComparison extends TACAS2015 {
 //		addTestCasesFixed("CodeCheckNoBE-C.xml", "TACASInterpolation2015/Kojak-Princess.epf", testcases);		
 	}	
 	
+	/**
+	 * {@inheritDoc}
+	 */
 	@Override
-	protected int getTimeout() {
+	protected long getTimeout() {
 		return 60 * 1000;
 	}
